refactor: move usage notice logic to TypeScript
